feat: handle http request as JsValues from Rust
When an HTTP request is received, the `Request` object is sanitized
and turned into a naive implementation of the HTTP request.

Then is consumed by Rust via the "wasm_bindgen" where the object is
treated as a JsValue and deserialized using serde.

At this point, the data incomming is consumed by Rust and operations
are performed from this side now on.